Added validation for device save/update from edge
